chore: Update Admin.scss and ManagerUser.js
This commit updates the Admin.scss file by removing the border property from the .admin-header class. It also comments out the .admin-main class in the same file.

In the ManagerUser.js file, it imports the TableUser component and adds it to the JSX code. It also removes the unused import of ModalCreateUser.

The changes in this commit improve the styling of the admin header and add a table to display user data in the ManagerUser component.