feat: Improve user management by adding delete functionality
The code changes in `ManagerUser.js` and `ModalDeleteUser.js` introduce the ability to delete users. The `handleClickDeleteUser` function in `ManagerUser.js` is modified to set the `showModalDeleteUser` state to true and pass the user data to the `ModalDeleteUser` component. In `ModalDeleteUser.js`, the `handleSubmitDeleteUser` function is implemented to send a DELETE request to the API endpoint `/api/v1/participant` with the user ID. After a successful deletion, the user list is fetched again to reflect the updated data. This commit enhances the user management feature by providing the functionality to delete users.